Add prompt injection mitigations for external content
Fence tool outputs from network-fetching tools (browser, web_fetch,
http_request, gmail, etc.) with adversarial-content markers at the
runner level so the model treats them as data, not instructions.
Extract html_to_text into a shared sanitize module, fence recalled
memories in the system prompt, and add a security policy section
that instructs the model to never follow directives found in
external content.
